# docs/swagger.yaml
basePath: /api
definitions:
  common.APIResponse:
    description: 所有 API 响应的统一包装格式，包含状态码、消息、数据和 timestamp
    properties:
      code:
        description: 状态码，200 表示成功，其他表示错误
        example: 200
        type: integer
      data:
        description: 响应数据，成功时返回业务数据，错误时为 null
      message:
        description: 消息提示，成功时为"success"，错误时为错误描述
        example: success
        type: string
      timestamp:
        description: Unix 时间戳（秒）
        example: 1717200000
        type: integer
    type: object
  controllers.CreateLanguageTextRequest:
    description: 创建多语言文本的请求参数
    properties:
      description:
        description: 描述说明
        example: 首页欢迎语
        type: string
      enText:
        description: 英文文本
        example: Welcome
        type: string
      key:
        description: 文本键（唯一）
        example: home.welcome
        type: string
      module:
        description: 所属模块
        example: home
        type: string
      zhText:
        description: 中文文本
        example: 欢迎
        type: string
    required:
    - key
    - module
    type: object
...
  controllers.UpdateLanguageTextRequest:
    description: 更新多语言文本的请求参数
    properties:
      description:
        description: 描述说明
        example: 首页欢迎语
        type: string
      enText:
        description: 英文文本
        example: Welcome
        type: string
      zhText:
        description: 中文文本
        example: 欢迎
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.LanguageText:
    description: 存储系统多语言文本，支持版本管理
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 描述说明
        example: 首页欢迎语
        type: string
      enText:
        description: 英文文本
        example: Welcome
        type: string
      id:
        type: integer
      key:
        description: 文本键（唯一）
        example: home.welcome
        type: string
      module:
        description: 所属模块
        example: home
        type: string
      updatedAt:
        type: string
      version:
        description: 版本号
        example: 1
        type: integer
      zhText:
        description: 中文文本
        example: 欢迎
        type: string
    type: object
  models.LanguageTextVersion:
    description: 存储多语言文本的历史版本，用于版本回溯
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 描述说明
        example: 首页欢迎语
        type: string
      enText:
        description: 英文文本
        example: Welcome
        type: string
      id:
        type: integer
      key:
        description: 文本键
        example: home.welcome
        type: string
      languageTextId:
        description: 关联的 LanguageText ID
        example: 1
        type: integer
      module:
        description: 所属模块
        example: home
        type: string
      updatedAt:
        description: 更新时间
        example: "2024-01-01T00:00:00Z"
        type: string
      version:
        description: 版本号
        example: 1
        type: integer
      zhText:
        description: 中文文本
        example: 欢迎
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    医疗器械内容管理系统 API 接口文档
    提供内容管理、模块配置、图片管理、多语言支持等功能
  title: Shirley CMS API
  version: "1.0"
paths:
...
  /api/admin/lang:
    get:
      consumes:
      - application/json
      description: 获取指定模块或所有模块的多语言文本列表
      parameters:
      - description: 模块名称，不传则返回所有模块
        in: query
        name: module
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 多语言文本列表
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.LanguageText'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 获取多语言文本列表
      tags:
      - 多语言管理
    post:
      consumes:
      - application/json
      description: 创建新的多语言文本条目
      parameters:
      - description: 多语言文本创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateLanguageTextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.LanguageText'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.APIResponse'
        "409":
          description: 键已存在
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 创建多语言文本
      tags:
      - 多语言管理
  /api/admin/lang/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的多语言文本
      parameters:
      - description: 多语言文本 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: 多语言文本不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 删除多语言文本
      tags:
      - 多语言管理
    get:
      consumes:
      - application/json
      description: 根据 ID 获取多语言文本详情
      parameters:
      - description: 多语言文本 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 多语言文本详情
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.LanguageText'
              type: object
        "404":
          description: 多语言文本不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 获取单个多语言文本
      tags:
      - 多语言管理
    put:
      consumes:
      - application/json
      description: 更新指定多语言文本的内容
      parameters:
      - description: 多语言文本 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 多语言文本更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateLanguageTextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.LanguageText'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: 多语言文本不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 更新多语言文本
      tags:
      - 多语言管理
  /api/admin/lang/{id}/restore/{version}:
    post:
      consumes:
      - application/json
      description: 将指定多语言文本恢复到某个历史版本
      parameters:
      - description: 多语言文本 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 版本号
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 恢复成功
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.LanguageText'
              type: object
        "404":
          description: 版本不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 恢复多语言文本版本
      tags:
      - 多语言管理
  /api/admin/lang/{id}/versions:
    get:
      consumes:
      - application/json
      description: 获取指定多语言文本的所有历史版本
      parameters:
      - description: 多语言文本 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 版本历史列表
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.LanguageTextVersion'
                  type: array
              type: object
        "404":
          description: 多语言文本不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 获取多语言文本版本历史
      tags:
      - 多语言管理
...
  /api/public/lang:
    get:
      consumes:
      - application/json
      description: 获取指定模块或所有模块的公开多语言文本（无需认证）
      parameters:
      - description: 模块名称，不传则返回所有模块
        in: query
        name: module
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 多语言文本列表
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.LanguageText'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: 获取公开多语言文本
      tags:
      - 公开接口
...
securityDefinitions:
  APIKey:
    description: JWT Token 认证，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"